batch: Move line range views into package
The line-matching package contains sequence operations while indexed views
over matched ranges remain at the batch package root.

Realization and baseline correspondence use those views to inspect matched
content, making the flat location disconnected from the responsibility it
supports.

This commit moves the range-view implementation into line_matching while
maintaining the selected import path during test migration.

The next source commit will move source-line lineage into the package.
